refactor: ApiResponse 타입 경로 수정 및 불필요한 파일 삭제
