+++ frontend/src/apis/notifications.api.ts
@@ -1,5 +1,6 @@
 import { apiClient } from '@/apis/apiClient';
-import { ApiResponse } from '@/types/api.types';
+import { ApiResponse } from '@/types/apiResponse';
+
 import type {
   FCMTokenRequest,
   NotificationSettingRequest,
