--- frontend/src/types/api.types.ts
@@ -1,5 +1,0 @@
-export interface ApiResponse<T = null> {
-  data: T;
-  status: number;
-  message: string;
-}
