+++ frontend/src/types/apiResponse.ts
@@ -1,4 +1,4 @@
-export interface ApiResponse<T> {
+export interface ApiResponse<T = null> {
   data: T;
   status?: number;
   message?: string;
